+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 07/xx/2017
 
 1. [](#new)
-    * Added a new `cache-control` system and page level property [#1591](https://github.com/getgrav/grav/issues/1591)
+    * Added a new `cache_control` system and page level property [#1591](https://github.com/getgrav/grav/issues/1591)
+    * Added a new `clear_images_by_default` system property to stop cache clear events from removing processed images [#1481](https://github.com/getgrav/grav/pull/1481)
 
 # v1.3.1
 ## 07/19/2017