+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 1. [](#new)
     * Added a new `cache_control` system and page level property [#1591](https://github.com/getgrav/grav/issues/1591)
     * Added a new `clear_images_by_default` system property to stop cache clear events from removing processed images [#1481](https://github.com/getgrav/grav/pull/1481)
-
+1. [](#bugfix)
+    * Allow `session.timetout` field to be set to `0` via blueprints [#1598](https://github.com/getgrav/grav/issues/1598)
+    
 # v1.3.1
 ## 07/19/2017